Extract `ClientOptions` to helper class
